Fix pep8
